dont restrict sub ids to uint16_t, throw out_of_bounds if string IDs are longer than numeric_limits<uint16_t>::max(), enable narrowing-conversion warnings again, should fix the issues raised in PR #17

+++ src/spatialjoin/GeometryCache.h
@@ -36,7 +36,7 @@ struct Area {
   std::string id;
 
   // sub id (for multipolygons)
-  uint16_t subId;
+  size_t subId;
 
   // area
   double area;
@@ -85,7 +85,7 @@ struct Line {
   std::string id;
 
   // sub id (for multilines)
-  uint16_t subId;
+  size_t subId;
 
   // length
   double length;
@@ -102,7 +102,7 @@ struct Point {
   std::string id;
 
   // sub id (for multipoints)
-  uint16_t subId;
+  size_t subId;
 };
 
 template <typename W>
